Added ability to clip with percentage ranges. Adapted two thumb slider to handle percentage ranges. Rounded edges and incorperated styling similar to Discord.

+++ src/timestamps.py
@@ -1,7 +1,7 @@
 #!/bin/python
 import time
 
-def from_seconds(seconds: int) -> time:
+def from_seconds(seconds: float) -> time:
     return time.gmtime(seconds)
